refactor: Search API 라우트 핸들러를 UseCase 패턴에 맞게 수정
